sign off matty

+++ consumer-site-plugin/src/main/java/com/photon/phresco/convertor/CsvXmlConvertor.java
@@ -8,7 +8,9 @@ import java.io.FileReader;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.Iterator;
 import java.util.List;
+import java.util.Map;
 import java.util.Set;
 
 import org.apache.maven.plugin.MojoExecutionException;
@@ -23,6 +25,7 @@ import sun.util.logging.resources.logging;
 import au.com.bytecode.opencsv.CSVReader;
 
 import com.jcraft.jsch.Buffer;
+import com.photon.phresco.exception.PhrescoException;
 import com.photon.phresco.parser.LocaleExtractor;
 import com.photon.phresco.plugin.commons.MavenProjectInfo;
 import com.photon.phresco.status.ConversionStatus;
@@ -30,6 +33,9 @@ import com.photon.phresco.status.ValidationStatus;
 import com.photon.phresco.validator.IValidator;
 import com.photon.phresco.vo.ColumnVO;
 import com.photon.phresco.vo.CsvFileVO;
+import com.photon.phresco.vo.ImageVO;
+
+import edu.emory.mathcs.backport.java.util.Arrays;
 
 public class CsvXmlConvertor {
 
@@ -42,9 +48,10 @@ public class CsvXmlConvertor {
 
 	public CsvXmlConvertor(MavenProjectInfo mavenProjectInfo, String manifestFileName, String phrescoTargetDir)
 			throws IOException, Exception {
-		System.setProperty("javax.xml.parsers.SAXParserFactory", "com.sun.org.apache.xerces.internal.jaxp.SAXParserFactoryImpl");
+		System.setProperty("javax.xml.parsers.SAXParserFactory",
+				"com.sun.org.apache.xerces.internal.jaxp.SAXParserFactoryImpl");
 		builder = new SAXBuilder();
-		
+
 		// disabling xml validation
 		builder.setValidation(false);
 		builder.setIgnoringElementContentWhitespace(true);
@@ -56,8 +63,7 @@ public class CsvXmlConvertor {
 
 	}
 
-	public List<CsvFileVO> convert(MavenProjectInfo mavenProjectInfo) throws MojoExecutionException, JDOMException,
-			IOException, Exception {
+	public List<CsvFileVO> convert(MavenProjectInfo mavenProjectInfo) throws JDOMException, IOException, Exception {
 		List<Element> structureList = ((List<Element>) XPath.selectNodes(doc, "//validations/structure"));
 		List<CsvFileVO> fileVOList = new ArrayList<CsvFileVO>();
 		for (Element structure : structureList) {
@@ -78,12 +84,13 @@ public class CsvXmlConvertor {
 				}
 			}
 		}
+		transferFiles(fileVOList, mavenProjectInfo);
 		new PHPCreator().createPHPFile(fileVOList, mavenProjectInfo);
 		return fileVOList;
 	}
 
 	private List<CsvFileVO> getXMLData(Element file, String stringFile, String parentFolder, String fileType,
-			String currentFolder) throws IOException {
+			String currentFolder) throws IOException, PhrescoException {
 		List<CsvFileVO> fileVOList = new ArrayList<CsvFileVO>();
 		if (stringFile.indexOf('!') >= 0) {
 			String replacementStringFile = "";
@@ -105,12 +112,17 @@ public class CsvXmlConvertor {
 	}
 
 	private List<CsvFileVO> getCSVData(Element file, List<Element> columnList, File csvFile, String parentFile,
-			String currentFolder) throws IOException {
+			String currentFolder) throws IOException, PhrescoException {
 		List<String[]> allCsvStringValues = getAllValuesFromCSVFile(csvFile);
 		HashMap<String, ColumnVO> columnVOSet = getColumnData(columnList);
 		String headerArr[] = null;
 		List<CsvFileVO> fileVOList = new ArrayList<CsvFileVO>();
 		headerArr = allCsvStringValues.get(0);
+		int minCount = Integer.parseInt(file.getAttributeValue("minCount"));
+		String fileName = file.getAttributeValue("name");
+		if (minCount > 0 && allCsvStringValues.size() <= 1) {
+			throw new PhrescoException("File " + fileName + " needs to contain at least one line");
+		}
 		for (int i = 1; i < allCsvStringValues.size(); i++) {
 			String arr[] = allCsvStringValues.get(i);
 			CsvFileVO csvFileVO = new CsvFileVO();
@@ -121,12 +133,29 @@ public class CsvXmlConvertor {
 				csvFileVO.setPhpFunction(file.getAttributeValue("addFunction"));
 			}
 			for (int j = 0; j < arr.length; j++) {
-				ColumnVO columnVO = columnVOSet.get(headerArr[j]);
+				ColumnVO columnVO = columnVOSet.get(headerArr[j].trim());
+				Set<String> keys = columnVOSet.keySet();
+				boolean flag = true;
+				for (String k : keys) {
+					if(!Arrays.asList(headerArr).contains(k)){
+						flag = false;
+					}
+				}
+				if(flag==false){
+					throw new PhrescoException("The column names in file " + csvFile.getAbsolutePath()
+							+ " are not as per manifest file -");
+				}
+				if (columnVO == null) {
+					throw new PhrescoException("The column names in file " + csvFile.getAbsolutePath()
+							+ " are not as per manifest file +");
+				}
 				String fieldName = columnVO.getFieldName();
 				String fieldType = columnVO.getFieldType();
+				if (!columnVO.isEmpty() && (arr[j] == null || arr[j].length() <= 0)) {
+					throw new PhrescoException("Column " + fieldName + " cannot be empty for file " + fileName);
+				}
 				csvFileVO.setFieldType(fieldType);
 				csvFileVO.setCurrentFolderPath(currentFolder.replace("!languages", "en"));
-				csvFileVO.setTargetFolder(columnVO.getDestination());
 				if (fieldName.equals("title") && fieldType.equals("argument")
 						&& (arr[j] != null && arr[j].length() > 0)) {
 					csvFileVO.getTitleMap().put("title", arr[j]);
@@ -142,7 +171,18 @@ public class CsvXmlConvertor {
 							csvFileVO.getExtraMap().put(fieldName, arr[j]);
 						}
 						if (fieldType.equals("image")) {
-							csvFileVO.getImageMap().put(fieldName, arr[j]);
+							ImageVO imageVO = new ImageVO();
+							imageVO.setFileName(arr[j]);
+							imageVO.setDestination(columnVO.getDestination());
+							imageVO.setKey(fieldName);
+							imageVO.setFileNameSize(columnVO.getFileNameSize());
+							imageVO.setDatatype(columnVO.getDatatype());
+							csvFileVO.getImageMap().put(fieldName, imageVO);
+							if (arr[j].length() > (columnVO.getFileNameSize())
+									&& columnVO.getDatatype().equals("varchar")) {
+								throw new PhrescoException("File Size not as per specifications for " + fileName
+										+ " image " + arr[j]);
+							}
 						}
 						if (fieldType.equals("metadata")) {
 							csvFileVO.getMetadataMap().put(fieldName, arr[j]);
@@ -161,11 +201,18 @@ public class CsvXmlConvertor {
 	private HashMap<String, ColumnVO> getColumnData(List<Element> columnList) {
 		HashMap<String, ColumnVO> columnVOSet = new HashMap<String, ColumnVO>();
 		for (Element column : columnList) {
-			String columnName = column.getAttribute("name").getValue();
+			String columnName = column.getAttribute("name").getValue().trim();
 			ColumnVO columnVO = new ColumnVO();
 			columnVO.setFieldName(column.getAttribute("fieldname").getValue());
 			columnVO.setFieldType(column.getAttribute("fieldtype").getValue());
-			columnVO.setDestination(column.getChild("destination").getValue());
+			if (column.getAttribute("fieldtype").getValue().equals("image")) {
+				columnVO.setDestination(column.getChild("destination").getValue());
+			}
+			columnVO.setEmpty(new Boolean(column.getAttributeValue("isEmpty")));
+			if (column.getChild("limit").getValue() != null && column.getChild("limit").getValue().length() > 0) {
+				columnVO.setFileNameSize(Integer.parseInt(column.getChild("limit").getValue()));
+			}
+			columnVO.setDatatype(column.getChild("datatype").getValue());
 			columnVOSet.put(columnName, columnVO);
 		}
 		return columnVOSet;
@@ -177,35 +224,38 @@ public class CsvXmlConvertor {
 	}
 
 	public void transferFiles(List<CsvFileVO> fileList, MavenProjectInfo mavenProjectInfo)
-			throws MojoExecutionException, JDOMException, IOException {
+			throws MojoExecutionException, JDOMException, IOException, PhrescoException {
 		for (CsvFileVO csvoFile : fileList) {
 			Set<String> keys = csvoFile.getImageMap().keySet();
-			String imageFileName = "";
 			String destination = "";
 			String source = "";
+			ImageVO imageVO = new ImageVO();
 			for (String k : keys) {
-				if (csvoFile.getFieldType().equals("image")) {
-					imageFileName = csvoFile.getImageMap().get(k);
-					source = mavenProjectInfo.getProject().getBasedir()
-							+ mavenProjectInfo.getProject().getProperties().getProperty("phresco.content.target.dir")
-							+ csvoFile.getCurrentFolderPath() + File.separator + "assets" + File.separator + "images"
-							+ File.separator + imageFileName;
-					destination = mavenProjectInfo.getProject().getBasedir() + File.separator + "source"
-							+ File.separator + csvoFile.getTargetFolder() + imageFileName;
-					File sourceFile = new File(source);
-					File destinationFile = new File(destination);
-					System.out.println("source  :  " + source);
-					System.out.println();
-					System.out.println("destination :  " + destination);
-					transferFile(mavenProjectInfo.getProject().getBasedir() + File.separator + "source"
-							+ File.separator + csvoFile.getTargetFolder(), sourceFile, destinationFile);
-				}
+				imageVO = csvoFile.getImageMap().get(k);
+				source = mavenProjectInfo.getProject().getBasedir()
+						+ mavenProjectInfo.getProject().getProperties().getProperty("phresco.content.target.dir")
+						+ csvoFile.getCurrentFolderPath() + File.separator + "assets" + File.separator + "images"
+						+ File.separator + imageVO.getFileName();
+				destination = mavenProjectInfo.getProject().getBasedir() + File.separator + "source" + File.separator
+						+ csvoFile.getImageMap().get(k).getDestination() + imageVO.getFileName();
+				System.out.println("csvoFile.getImageMap().get(k).getDestination()   "
+						+ csvoFile.getImageMap().get(k).getDestination());
+				File sourceFile = new File(source);
+				File destinationFile = new File(destination);
+				System.out.println("source  :  " + source);
+				System.out.println();
+				System.out.println("destination :  " + destination);
+				transferFile(mavenProjectInfo.getProject().getBasedir() + File.separator + "source" + File.separator
+						+ csvoFile.getImageMap().get(k).getDestination(), sourceFile, destinationFile);
 			}
 		}
 	}
 
 	private List<ConversionStatus> transferFile(String destinationDir, File sourceFile, File destinationFile)
-			throws MojoExecutionException, IOException {
+			throws MojoExecutionException, IOException, PhrescoException {
+		if (!sourceFile.exists()) {
+			throw new PhrescoException("File " + sourceFile.getAbsolutePath() + " not found");
+		}
 		new File(destinationDir).mkdirs();
 		FileInputStream fileInputStream = new FileInputStream(sourceFile);
 		FileOutputStream fileOutputStream = new FileOutputStream(destinationFile);
